@@ -2,7 +2,6 @@ package sawyern.cookiebot.bot;
 
 import discord4j.core.DiscordClient;
 import discord4j.core.event.domain.message.MessageCreateEvent;
-import discord4j.core.object.entity.Message;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import sawyern.cookiebot.exception.InvalidMessageCookieException;
@@ -14,6 +13,7 @@ import sawyern.cookiebot.util.BotUtil;
 
 import java.text.MessageFormat;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -21,7 +21,7 @@ import java.util.regex.Pattern;
 public abstract class GenericBotCommand implements BotCommand {
     private List<String> args;
 
-    private static Logger LOGGER = LoggerFactory.getLogger(GenericBotCommand.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(GenericBotCommand.class);
 
     /**
      * subscribes this command to be executed when the input message starts with "!" and matches getCommand()
@@ -36,7 +36,7 @@ public abstract class GenericBotCommand implements BotCommand {
         try {
             String content = event.getMessage().getContent().orElseThrow(DiscordException::new);
             if (parseCommand(content)) {
-                LOGGER.info(MessageFormat.format("Received command: {0}", getCommand()));
+                LOGGER.info("Received command: {}", getCommand());
 
                 // execute command
                 execute(event);
@@ -46,7 +46,6 @@ public abstract class GenericBotCommand implements BotCommand {
             BotUtil.sendMessage(event, e.getMessage());
         } catch (DiscordException e) {
             // text message was not sent. could be image etc. ignore this case
-            return;
         } catch (Exception e) {
             // log unknown exceptions
             LOGGER.error(e.getMessage(), e);
@@ -54,10 +53,27 @@ public abstract class GenericBotCommand implements BotCommand {
     }
 
     /**
-     * name of command
      * @return the name of the command
+     *
+     * ex. Input: !myCommand arg1 arg2
+     *     getCommand() returns "myCommand"
      */
     public abstract String getCommand();
+
+    /**
+     * default is 0, override to change. Must not return null
+     * @return a list of acceptable argument lengths
+     * @see #checkArgs()
+     */
+    public List<Integer> allowedNumArgs() {
+        //default number of arguments is 0
+        return Collections.singletonList(0);
+    }
+
+    protected final void checkArgs() throws CookieException {
+        if (!allowedNumArgs().contains(getArgs().size()))
+            throw new CookieException(MessageFormat.format("Invalid number of arguments. Expected {0}", allowedNumArgs()));
+    }
 
     /**
      * returns true if input string starts with "!" and matches getCommand()
@@ -68,7 +84,7 @@ public abstract class GenericBotCommand implements BotCommand {
     public boolean parseCommand(String message) throws CookieException {
         if (message == null || !message.startsWith(CommandConstants.COMMAND_START))
             return false;
-        this.args = parseArgs(message);
+        parseArgs(message);
         String command = args.stream().findFirst().orElseThrow(MessageParseCommandCookieException::new);
         return command.substring(1).equalsIgnoreCase(getCommand());
     }
@@ -76,20 +92,24 @@ public abstract class GenericBotCommand implements BotCommand {
     /**
      * splits the message by " " into a list of string arguments.
      * if argument is surrounded by quotes, it will consider the block as one argument
+     * sets local msgArgs variable
      * @param message input message
-     * @return list of arguments
      */
-    private List<String> parseArgs(String message) throws CookieException {
+    private void parseArgs(String message) throws CookieException {
         if (message == null || message.isEmpty())
             throw new InvalidMessageCookieException();
         List<String> msgArgs = new ArrayList<>();
         Matcher m = Pattern.compile(CommandConstants.QUOTE_REGEX).matcher(message);
         while (m.find())
             msgArgs.add(m.group(1).replace(CommandConstants.QUOTE, CommandConstants.EMPTY_STRING));
-        return msgArgs;
+        this.args = msgArgs;
+        checkArgs();
     }
 
-    public List<String> getArgs() {
+    public List<String> getArgs() throws CookieException {
+        if (this.args == null)
+            throw new CookieException("Args not yet parsed.");
         return args;
     }
+
 }
