@@ -1,41 +1,23 @@
 package sawyern.cookiebot.models.entity;
 
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NonNull;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
 
 @Entity
 @Table(name = "ACCOUNT")
+@Data
+@EqualsAndHashCode(callSuper = true)
 public class Account extends DbItem {
     @Column(name = "discordId")
+    @NonNull
     private String discordId;
 
     @Column(name = "username")
+    @NonNull
     private String username;
-
-    public Account() {
-        super();
-    }
-
-    public Account(String discordId, String username) {
-        super();
-        this.discordId = discordId;
-        this.username = username;
-    }
-
-    public String getDiscordId() {
-        return discordId;
-    }
-
-    public void setDiscordId(String discordId) {
-        this.discordId = discordId;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
 }
