@@ -1,19 +1,16 @@
 package sawyern.cookiebot.models.entity;
 
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
 import javax.persistence.*;
 
 @Entity
 @Table(name = "LOOTBOX_TOKEN")
+@Data
+@EqualsAndHashCode(callSuper = true)
 public class LootboxToken extends DbItem {
     @ManyToOne
     @JoinColumn(name = "ACCOUNT_ID")
     private Account account;
-
-    public Account getAccount() {
-        return account;
-    }
-
-    public void setAccount(Account account) {
-        this.account = account;
-    }
 }
