@@ -1,10 +1,15 @@
 package sawyern.cookiebot.models.entity;
 
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
 import javax.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "WEEKLY_COOLDOWN")
+@Data
+@EqualsAndHashCode(callSuper = true)
 public class WeeklyCooldown extends DbItem {
     @OneToOne
     @JoinColumn(name = "ACCOUNT_ID")
@@ -12,23 +17,4 @@ public class WeeklyCooldown extends DbItem {
 
     @Column
     private LocalDateTime canUseNext;
-
-    public WeeklyCooldown() {
-    }
-
-    public Account getAccount() {
-        return account;
-    }
-
-    public void setAccount(Account account) {
-        this.account = account;
-    }
-
-    public LocalDateTime getCanUseNext() {
-        return canUseNext;
-    }
-
-    public void setCanUseNext(LocalDateTime canUseNext) {
-        this.canUseNext = canUseNext;
-    }
 }
